refactor: postSticker 분리
